backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
beanie==1.23.6  # ODM for MongoDB

# AI and ML
google-generativeai==0.8.5
openai==1.3.7
anthropic==0.7.8

# Image processing
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-image==0.21.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching and performance
redis==5.0.1
aioredis==2.0.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
